serviceRecordHandle is long in java
git-svn-id: http://bluecove.googlecode.com/svn/trunk@930 3cac39a0-5d1b-0410-8711-c1891c4e0f47

+++ src/main/java/com/intel/bluetooth/BluetoothStackBlueSoleil.java
@@ -224,7 +224,7 @@ class BluetoothStackBlueSoleil implements BluetoothStack {
 	}
 	 */
 
-	public void servicesFoundCallback(SearchServicesThread startedNotify, DiscoveryListener listener, RemoteDevice device, String serviceName, byte[] uuidValue, int channel, int recordHanlde) {
+	public void servicesFoundCallback(SearchServicesThread startedNotify, DiscoveryListener listener, RemoteDevice device, String serviceName, byte[] uuidValue, int channel, long recordHanlde) {
 		ServiceRecordImpl record = new ServiceRecordImpl(device, 0);
 
 		UUID uuid = new UUID(Utils.UUIDByteArrayToString(uuidValue), false);
